Rust - split example 1 into 3 parts
