+++ examples/rust/ex1-volume/src/main.rs
@@ -30,77 +30,20 @@
 use libceed::{prelude::*, Ceed};
 use structopt::StructOpt;
 
-// ----------------------------------------------------------------------------
-// Command line arguments
-// ----------------------------------------------------------------------------
-#[derive(Debug, StructOpt)]
-#[structopt(
-    name = "libCEED Rust Example 1 - Volume",
-    about = "This example uses the mass matrix to compute the length, area, or volume of a region, depending upon runtime parameters."
-)]
-struct Opt {
-    /// libCEED backend resource to use
-    #[structopt(name = "ceed", short = "c", long = "ceed", default_value = "/cpu/self")]
-    ceed_spec: String,
-    /// Mesh dimension
-    #[structopt(
-        name = "dimension",
-        short = "d",
-        long = "dimension",
-        default_value = "3"
-    )]
-    dim: usize,
-    /// Polynomial degree for the mesh
-    #[structopt(
-        name = "mesh degree",
-        short = "m",
-        long = "mesh_degree",
-        default_value = "4"
-    )]
-    mesh_degree: usize,
-    /// Polynomial degree for the solution
-    #[structopt(
-        name = "solution degree",
-        short = "p",
-        long = "solution_degree",
-        default_value = "4"
-    )]
-    solution_degree: usize,
-    /// Number of quadrature points in 1D
-    #[structopt(
-        name = "number of quadrature points",
-        short = "q",
-        long = "num_qpts",
-        default_value = "6"
-    )]
-    num_qpts: usize,
-    /// Approximate problem size
-    #[structopt(
-        name = "problem size",
-        short = "s",
-        long = "problem_size",
-        default_value = "-1"
-    )]
-    problem_size_requested: i64,
-    /// Test mode
-    #[structopt(name = "test mode", short = "t", long = "test")]
-    test: bool,
-    /// Gallery QFunctions
-    #[structopt(name = "gallery QFunctions", short = "g", long = "gallery")]
-    gallery: bool,
-}
+mod mesh;
+mod opt;
 
 // ----------------------------------------------------------------------------
 // Example 1
 // ----------------------------------------------------------------------------
 fn main() -> Result<(), String> {
-    let opt = Opt::from_args();
-    example_1(opt)
+    let options = opt::Opt::from_args();
+    example_1(options)
 }
 
-fn example_1(opt: Opt) -> Result<(), String> {
+fn example_1(options: opt::Opt) -> Result<(), String> {
     // Process command line arguments
-    let Opt {
+    let opt::Opt {
         ceed_spec,
         dim,
         mesh_degree,
@@ -109,7 +52,7 @@ fn example_1(opt: Opt) -> Result<(), String> {
         problem_size_requested,
         test,
         gallery,
-    } = opt;
+    } = options;
     assert!(dim >= 1 && dim <= 3);
     assert!(mesh_degree >= 1);
     assert!(solution_degree >= 1);
@@ -147,7 +90,7 @@ fn example_1(opt: Opt) -> Result<(), String> {
         ceed.basis_tensor_H1_Lagrange(dim, 1, solution_degree + 1, num_qpts, QuadMode::Gauss);
 
     // Determine mesh size from approximate problem size
-    let num_xyz = get_cartesian_mesh_size(dim, solution_degree, problem_size);
+    let num_xyz = mesh::get_cartesian_mesh_size(dim, solution_degree, problem_size);
     if !test {
         print!("\nMesh size                   : nx = {}", num_xyz[0]);
         if dim > 1 {
@@ -162,9 +105,9 @@ fn example_1(opt: Opt) -> Result<(), String> {
     // Build ElemRestriction objects describing the mesh and solution discrete
     // representations
     let (restr_mesh, _) =
-        build_cartesian_restriction(&ceed, dim, num_xyz, mesh_degree, ncomp_x, num_qpts);
+        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, mesh_degree, ncomp_x, num_qpts);
     let (restr_solution, restr_qdata) =
-        build_cartesian_restriction(&ceed, dim, num_xyz, solution_degree, 1, num_qpts);
+        mesh::build_cartesian_restriction(&ceed, dim, num_xyz, solution_degree, 1, num_qpts);
     let mesh_size = restr_mesh.get_lvector_size() as usize;
     let solution_size = restr_solution.get_lvector_size() as usize;
     if !test {
@@ -173,10 +116,11 @@ fn example_1(opt: Opt) -> Result<(), String> {
     }
 
     // Create a Vector with the mesh coordinates
-    let mut mesh_coords = set_cartesian_mesh_coords(&ceed, dim, num_xyz, mesh_degree, mesh_size);
+    let mut mesh_coords =
+        mesh::set_cartesian_mesh_coords(&ceed, dim, num_xyz, mesh_degree, mesh_size);
 
     // Apply a transformation to the mesh coordinates
-    let exact_volume = transform_mesh_coordinates(dim, mesh_size, &mut mesh_coords);
+    let exact_volume = mesh::transform_mesh_coordinates(dim, mesh_size, &mut mesh_coords);
 
     // QFunction that builds the quadrature data for the mass operator
     // -- QFunction from user closure
@@ -340,184 +284,6 @@ fn example_1(opt: Opt) -> Result<(), String> {
 }
 
 // ----------------------------------------------------------------------------
-// Determine problem size in each dimension from size and dimenison
-// ----------------------------------------------------------------------------
-fn get_cartesian_mesh_size(dim: usize, solution_degree: usize, problem_size: i64) -> [usize; 3] {
-    // Use the approximate formula:
-    //    prob_size ~ num_elem * degree^dim
-    let mut num_elem = problem_size / solution_degree.pow(dim as u32) as i64;
-    let mut s = 0; // find s: num_elem / 2 < 2^s <= num_elem
-
-    while num_elem > 1 {
-        num_elem /= 2;
-        s += 1;
-    }
-
-    let mut r = s % dim;
-    let mut num_xyz = [0; 3];
-    for d in 0..dim {
-        let mut sd = s / dim;
-        if r > 0 {
-            sd += 1;
-            r -= 1;
-        }
-        num_xyz[d] = 1 << sd;
-    }
-    num_xyz
-}
-
-// ----------------------------------------------------------------------------
-// Build element restriction objects for the mesh
-// ----------------------------------------------------------------------------
-fn build_cartesian_restriction(
-    ceed: &Ceed,
-    dim: usize,
-    num_xyz: [usize; 3],
-    degree: usize,
-    num_comp: usize,
-    num_qpts: usize,
-) -> (ElemRestriction, ElemRestriction) {
-    let p = degree + 1;
-    let num_nodes = p.pow(dim as u32); // number of nodes per element
-    let elem_qpts = num_qpts.pow(dim as u32); // number of quadrature pts per element
-
-    let mut num_d = [0; 3];
-    let mut num_elem = 1;
-    let mut scalar_size = 1;
-    for d in 0..dim {
-        num_elem *= num_xyz[d];
-        num_d[d] = num_xyz[d] * (p - 1) + 1;
-        scalar_size *= num_d[d];
-    }
-
-    // elem:          0             1                 n-1
-    //         |---*-...-*---|---*-...-*---|- ... -|--...--|
-    // nnodes: 0   1    p-1  p  p+1       2*p             n*p
-    let mut elem_nodes = vec![0; num_elem * num_nodes];
-    for e in 0..num_elem {
-        let mut e_xyz = [1; 3];
-        let mut re = e;
-        for d in 0..dim {
-            e_xyz[d] = re % num_xyz[d];
-            re /= num_xyz[d];
-        }
-        let loc_offset = e * num_nodes;
-        for loc_nodes in 0..num_nodes {
-            let mut global_nodes = 0;
-            let mut global_nodes_stride = 1;
-            let mut r_nodes = loc_nodes;
-            for d in 0..dim {
-                global_nodes += (e_xyz[d] * (p - 1) + r_nodes % p) * global_nodes_stride;
-                global_nodes_stride *= num_d[d];
-                r_nodes /= p;
-            }
-            elem_nodes[loc_offset + loc_nodes] = global_nodes as i32;
-        }
-    }
-
-    // Mesh/solution data restriction
-    let restr = ceed.elem_restriction(
-        num_elem,
-        num_nodes,
-        num_comp,
-        scalar_size,
-        num_comp * scalar_size,
-        MemType::Host,
-        &elem_nodes,
-    );
-    // Quadratue data restriction
-    let strides: [i32; 3] = [1, elem_qpts as i32, elem_qpts as i32];
-    let restr_qdata = ceed.strided_elem_restriction(
-        num_elem,
-        elem_qpts,
-        num_comp,
-        num_comp * elem_qpts * num_elem,
-        strides,
-    );
-    (restr, restr_qdata)
-}
-
-// ----------------------------------------------------------------------------
-// Set mesh coordinates
-// ----------------------------------------------------------------------------
-fn set_cartesian_mesh_coords(
-    ceed: &Ceed,
-    dim: usize,
-    num_xyz: [usize; 3],
-    mesh_degree: usize,
-    mesh_size: usize,
-) -> Vector {
-    let p = mesh_degree + 1;
-    let mut num_d = [0; 3];
-    let mut scalar_size = 1;
-    for d in 0..dim {
-        num_d[d] = num_xyz[d] * (p - 1) + 1;
-        scalar_size *= num_d[d];
-    }
-
-    // Lobatto points
-    let lobatto_basis = ceed.basis_tensor_H1_Lagrange(1, 1, 2, p, QuadMode::GaussLobatto);
-    let nodes_corners = ceed.vector_from_slice(&[0.0, 1.0]);
-    let mut nodes_full = ceed.vector(p);
-    lobatto_basis.apply(
-        1,
-        TransposeMode::NoTranspose,
-        EvalMode::Interp,
-        &nodes_corners,
-        &mut nodes_full,
-    );
-
-    let mut mesh_coords = ceed.vector(mesh_size);
-    {
-        let mut coords = mesh_coords.view_mut();
-        let nodes = nodes_full.view();
-        for gs_nodes in 0..scalar_size {
-            let mut r_nodes = gs_nodes;
-            for d in 0..dim {
-                let d_1d = r_nodes % num_d[d];
-                coords[gs_nodes + scalar_size * d] =
-                    ((d_1d / (p - 1)) as f64 + nodes[d_1d % (p - 1)]) / num_xyz[d] as f64;
-                r_nodes /= num_d[d];
-            }
-        }
-    }
-    mesh_coords
-}
-
-// ----------------------------------------------------------------------------
-// Transform mesh coordinates
-// ----------------------------------------------------------------------------
-fn transform_mesh_coordinates(dim: usize, mesh_size: usize, mesh_coords: &mut Vector) -> f64 {
-    let exact_volume = if dim == 1 {
-        1.0
-    } else {
-        3.0 / 4.0 * std::f64::consts::PI
-    };
-    let mut coords = mesh_coords.view_mut();
-
-    if dim == 1 {
-        for i in 0..mesh_size {
-            // map [0,1] to [0,1] varying the mesh density
-            coords[i] = 0.5
-                + 1.0 / (3.0_f64).sqrt()
-                    * ((2.0 / 3.0) * std::f64::consts::PI * (coords[i] - 0.5)).sin();
-        }
-    } else {
-        let num_nodes = mesh_size / dim;
-        for i in 0..num_nodes {
-            // map (x,y) from [0,1]x[0,1] to the quarter annulus with polar
-            // coordinates, (r,phi) in [1,2]x[0,pi/2] with area = 3/4*pi
-            let u = 1.0 + coords[i];
-            let v = std::f64::consts::PI / 2.0 * coords[i + num_nodes];
-            coords[i] = u * v.cos();
-            coords[i + num_nodes] = u * v.sin();
-        }
-    }
-
-    exact_volume
-}
-
-// ----------------------------------------------------------------------------
 // Tests
 // ----------------------------------------------------------------------------
 #[cfg(test)]
@@ -534,7 +300,7 @@ mod tests {
         let problem_size_requested = -1;
         let test = true;
         let gallery = false;
-        let opt = Opt {
+        let options = opt::Opt {
             ceed_spec,
             dim,
             mesh_degree,
@@ -544,7 +310,7 @@ mod tests {
             test,
             gallery,
         };
-        assert!(example_1(opt).is_ok());
+        assert!(example_1(options).is_ok());
     }
 
     #[test]
@@ -557,7 +323,7 @@ mod tests {
         let problem_size_requested = -1;
         let test = true;
         let gallery = false;
-        let opt = Opt {
+        let options = opt::Opt {
             ceed_spec,
             dim,
             mesh_degree,
@@ -567,7 +333,7 @@ mod tests {
             test,
             gallery,
         };
-        assert!(example_1(opt).is_ok());
+        assert!(example_1(options).is_ok());
     }
 
     #[test]
@@ -580,7 +346,7 @@ mod tests {
         let problem_size_requested = -1;
         let test = true;
         let gallery = false;
-        let opt = Opt {
+        let options = opt::Opt {
             ceed_spec,
             dim,
             mesh_degree,
@@ -590,7 +356,7 @@ mod tests {
             test,
             gallery,
         };
-        assert!(example_1(opt).is_ok());
+        assert!(example_1(options).is_ok());
     }
 
     #[test]
@@ -603,7 +369,7 @@ mod tests {
         let problem_size_requested = -1;
         let test = true;
         let gallery = true;
-        let opt = Opt {
+        let options = opt::Opt {
             ceed_spec,
             dim,
             mesh_degree,
@@ -613,7 +379,7 @@ mod tests {
             test,
             gallery,
         };
-        assert!(example_1(opt).is_ok());
+        assert!(example_1(options).is_ok());
     }
 }
